confirmation before deleting , different base currency asset and leverage multiplier implemented but only on account page not on global portfolio

+++ src/components/sections/home/Orderstable.js
@@ -8,9 +8,10 @@ import Fuse from 'fuse.js';
 import './app.css'
 import {saveAsset} from "../../Utility/assets/saveAsset";
 import {connect} from "react-redux";
-import { assetsListUpdate, coingeckoCryptosListFetch} from "../../../reducers/actions";
+import {assetsListUpdate, coingeckoCryptosListFetch} from "../../../reducers/actions";
 import {formatter} from "../../Utility/functions";
 import {deleteAsset} from "../../Utility/assets/deleteAsset";
+import CoinGecko from "coingecko-api";
 
 var trend_1 = [1, 3, 5, 4, 3, 3, 4, 8, 5];
 var trend_2 = [3, 2, 5, 4, 8, 4, 5, 2, 3];
@@ -110,7 +111,13 @@ const optionsData = {
     }
 }
 
-function fuzzySearch(options,data) {
+async function getAssetInfo(id) {
+    const CoinGeckoClient = new CoinGecko();
+    let el = await CoinGeckoClient.coins.fetch(id)
+    return el.data.image.small
+}
+
+function fuzzySearch(options, data) {
     const fuse = new Fuse(data, {
         keys: ['name', 'groupName', 'items.name'],
         threshold: 0.3,
@@ -120,7 +127,6 @@ function fuzzySearch(options,data) {
         if (!value.length) {
             return options;
         }
-
         return fuse.search(value);
     };
 }
@@ -173,7 +179,14 @@ const Orderstable = (props) => {
         trendfive(),
     ])
     const [data, setData] = useState([]);
+    const [vsCurrencies, setVsCurrencies] = useState([
+        {name: "BTC", value: 'btc'},
+        {name: "ETH", value: 'eth'},
+        {name: "USD", value: 'usd'},
+        {name: "EUR", value: 'eur'},
+    ]);
     const [show1, setShow1] = useState(false);
+    const [currentAssetImage, setCurrentAssetImage] = useState('');
     const [newAsset, setNewAsset] = useState({
         buyDate: new Date().toDateInputValue(),
         name: '',
@@ -182,7 +195,10 @@ const Orderstable = (props) => {
         quantity: 0,
         unitPrice: 0,
         fees: 0,
+        leverage: 1,
         description: '',
+        vsCurrency: '',
+        image: ''
     });
 
 
@@ -216,11 +232,22 @@ const Orderstable = (props) => {
     const handleChange = (...args) => {
         console.log("ARGS:", args);
         var temp = newAsset
+        getAssetInfo(args[0]).then(res => {
+            setCurrentAssetImage(res)
+            temp.image = res
+        })
         temp.name = args[1].name
         temp.symbol = args[1].symbol
         temp.id = args[0]
         setNewAsset(temp)
     };
+
+    const handleChangeCurrency = (...args) => {
+        var temp = newAsset
+        temp.vsCurrency = args[0]
+        setNewAsset(temp)
+    };
+
 
     return (
         <div className="col-md-12">
@@ -261,24 +288,26 @@ const Orderstable = (props) => {
                                     <td>{Intl.DateTimeFormat('de-DE').format(new Date(item.buyDate))}</td>
                                     <td className="ms-crypto-amount"><img style={{width: '20px'}} src={item.image} alt=""/>{item.name}</td>
                                     <td>{item.quantity + ' ' + item.symbol.toUpperCase()}</td>
-                                    <td>{formatter.format(item.spentAmount)}</td>
-                                    <td>${item.unitPrice}</td>
-                                    <td>{formatter.format(item.usdPrice)}</td>
-                                    <td>1</td>
+                                    <td>{item.vsCurrency === 'usd' || item.vsCurrency === undefined ? formatter.format(item.spentAmount) : item.spentAmount + ' ' + item.vsCurrency.toUpperCase()}</td>
+                                    <td>{item.vsCurrency === 'usd' || item.vsCurrency === undefined ? formatter.format(item.unitPrice) : item.unitPrice + ' ' + item.vsCurrency.toUpperCase() }</td>
+                                    <td>{ item.vsCurrency === 'usd' || item.vsCurrency === undefined ? formatter.format(item.usdPrice) : item.basePrice + ' ' + item.vsCurrency.toUpperCase()}</td>
+                                    <td>{item.leverage ?? 1}</td>
                                     <td className="ms-trend"><LineChart data={options[2]} options={optionsData}/></td>
                                     <td className="ms-text-info">{formatter.format(item.currentAmount)}</td>
                                     <td className={item.performancePercentage > 0 ? 'text-success' : 'text-danger'}>{item.performancePercentage + '%'}</td>
-                                    <td className={item.performanceUsd > 0 ? 'text-success' : 'text-danger'}>{formatter.format(item.performanceUsd)}</td>
+                                    <td className={item.performanceUsd > 0 ? 'text-success' : 'text-danger'}>{ formatter.format(item.performanceUsd) }</td>
                                     <td><i className="text-danger flaticon-trash ms-delete-trigger" onClick={(e) => {
-                                        console.log('deleted');
-                                        deleteAsset(item.id).then(res => {
-                                            let temp = props.assets.filter(function (obj) {
-                                                return obj.id !== item.id;
-                                            });
-                                            console.log("temp", temp);
-                                            console.log("assets", props.assets);
-                                            props.assetsListUpdate(temp)
-                                        })
+                                        if (window.confirm('Etes vous sur de vouloir supprimer cet actif ?')) {
+                                            console.log('deleted');
+                                            deleteAsset(item.id).then(res => {
+                                                let temp = props.assets.filter(function (obj) {
+                                                    return obj.id !== item.id;
+                                                });
+                                                props.assetsListUpdate(temp)
+                                            })
+                                        } else {
+
+                                        }
                                     }}> </i></td>
                                 </tr>
                             )) : null}
@@ -290,7 +319,7 @@ const Orderstable = (props) => {
             <Modal show={show1} onHide={handleClose} aria-labelledby="contained-modal-title-vcenter"
                    centered>
                 <Modal.Header>
-                    <h3 className="modal-title has-icon ms-icon-round "><i className="flaticon-share bg-primary text-white"/> Ajouter un actif</h3>
+                    <h3 className="modal-title has-icon ms-icon-round "><img src={currentAssetImage} alt=""/> Ajouter un actif</h3>
                     <button type="button" className="close" onClick={handleClose}><span aria-hidden="true">×</span></button>
                 </Modal.Header>
                 <Form onSubmit={handleSubmit}>
@@ -300,7 +329,7 @@ const Orderstable = (props) => {
                             controlId="validationCustom02">
                             <Form.Row style={{marginTop: "1rem"}}>
                                 <div className="col">
-                                    <Form.Label>Begin des Events</Form.Label>
+                                    <Form.Label>Date d'achat</Form.Label>
                                     <InputGroup>
                                         <Form.Control
                                             required
@@ -313,17 +342,28 @@ const Orderstable = (props) => {
                                     </InputGroup>
                                 </div>
                             </Form.Row>
-                            <Form.Row className={"has-icon"}>
+                            <Form.Row className={"has-icon mb-4"}>
                                 <label>Nom de l'actif</label>
+                                {/*renderOption={renderFriend}*/}
                                 {data !== null ?
-                                    <SelectSearch filterOptions={(options) => fuzzySearch(options,data)} autoComplete={"on"} onChange={handleChange} search={true} options={data.slice(0,20)} value="sv" name="assets"
+                                    <SelectSearch filterOptions={(options) => fuzzySearch(options, data)} autoComplete={"on"} onChange={handleChange} search={true} options={data.slice(0, 20)}
+                                                  value="sv" name="assets"
                                                   placeholder="Choose your Asset"/> : null}
+                            </Form.Row>
+                            <Form.Row className={"has-icon"}>
+                                <label>Actif de base</label>
+                                {/*renderOption={renderFriend}*/}
+                                {vsCurrencies !== null ?
+                                    <SelectSearch filterOptions={(options) => fuzzySearch(options, vsCurrencies)} autoComplete={"on"} onChange={handleChangeCurrency} search={true}
+                                                  options={vsCurrencies}
+                                                  value="sv" name="vscurrencies"
+                                                  placeholder="Choose the base Asset"/> : null}
                             </Form.Row>
                             <Form.Row>
                                 <div className="col-md-6">
                                     <div className="ms-form-group has-icon">
                                         <label>Quantite</label>
-                                        <input type="text" value={newAsset.quantity} placeholder="0.00" className="form-control" name="quantity"
+                                        <input type="number" value={newAsset.quantity} placeholder="0.00" className="form-control" name="quantity"
                                                onChange={e => {
                                                    setNewAsset(prev => ({...prev, quantity: e.target.value}))
                                                }}
@@ -333,7 +373,7 @@ const Orderstable = (props) => {
                                 <div className="col-md-6">
                                     <div className="ms-form-group has-icon">
                                         <label>Prix unitaire</label>
-                                        <input value={newAsset.unitPrice} type="text" placeholder="0.00" className="form-control" name="price"
+                                        <input value={newAsset.unitPrice} type="number" placeholder="0.00" className="form-control" name="price"
                                                onChange={e => {
                                                    setNewAsset(prev => ({...prev, unitPrice: e.target.value}))
                                                }}
@@ -342,8 +382,16 @@ const Orderstable = (props) => {
                                 </div>
                             </Form.Row>
                             <Form.Row className={"mb-2"}>
+                                <label>Levier</label>
+                                <input value={newAsset.leverage} type="number" placeholder="0.00" className="form-control" name="leverage"
+                                       onChange={e => {
+                                           setNewAsset(prev => ({...prev, leverage: e.target.value}))
+                                       }}
+                                />
+                            </Form.Row>
+                            <Form.Row className={"mb-2"}>
                                 <label>Frais</label>
-                                <input value={newAsset.fees} type="text" placeholder="0.00" className="form-control" name="fees"
+                                <input value={newAsset.fees} type="number" placeholder="0.00" className="form-control" name="fees"
                                        onChange={e => {
                                            setNewAsset(prev => ({...prev, fees: e.target.value}))
                                        }}
